🎨 Extract modal logic into composable
