+++ composables/use-tts-try-modal.ts
@@ -1,10 +1,12 @@
 import { useStorage } from '@vueuse/core'
+import { TTSTryModal } from '#components'
 
 const TTS_TRY_OFFER_KEY = '3ook_tts_try_offer'
 const TTS_TRY_COOLDOWN_KEY = '3ook_tts_try_cooldown'
 
 export function useTTSTryModal() {
   const { user } = useUserSession()
+  const overlay = useOverlay()
 
   const hasTTSTrialOffered = useStorage(TTS_TRY_OFFER_KEY, true)
 
@@ -24,13 +26,50 @@ export function useTTSTryModal() {
     hasTTSTrialOffered.value = false
   }
 
-  function snoozeTTSTryModal(durationMs: number) {
-    offerCooldown.value = Date.now() + durationMs
+  function snoozeTTSTryModal() {
+    const oneWeekInMs = 7 * 24 * 60 * 60 * 1000
+    offerCooldown.value = Date.now() + oneWeekInMs
   }
 
   function resetTTSTryOffer() {
     hasTTSTrialOffered.value = true
     offerCooldown.value = Date.now()
+  }
+
+  function showTTSTryModal(options: {
+    nftClassId: string
+    onVoiceSelected?: (languageVoice: string) => void
+  }) {
+    const modal = overlay.create(TTSTryModal, {
+      props: {
+        nftClassId: options.nftClassId,
+        onVoiceSelected: (languageVoice: string) => {
+          dismissTTSTryModal()
+          useLogEvent('tts_try_voice_selected', {
+            nft_class_id: options.nftClassId,
+            languageVoice,
+          })
+          modal.close()
+          options.onVoiceSelected?.(languageVoice)
+        },
+        onSnooze: () => {
+          snoozeTTSTryModal()
+          useLogEvent('tts_try_snoozed', {
+            nft_class_id: options.nftClassId,
+          })
+          modal.close()
+        },
+        onDismiss: () => {
+          dismissTTSTryModal()
+          useLogEvent('tts_try_dismissed', {
+            nft_class_id: options.nftClassId,
+          })
+          modal.close()
+        },
+      },
+    })
+
+    modal.open()
   }
 
   return {
@@ -40,5 +79,6 @@ export function useTTSTryModal() {
     dismissTTSTryModal,
     snoozeTTSTryModal,
     resetTTSTryOffer,
+    showTTSTryModal,
   }
 }
